wiserylabs/app/requirements.txt
fastapi
uvicorn
python-docx
openai
langchain
langchain-openai
